Refactor dump_txout_set
