+++ crates/subcoin-node/src/commands/blockchain/dump_txout_set.rs
@@ -18,15 +18,13 @@ pub struct DumpTxOutSet {
     #[clap(long)]
     height: Option<u32>,
 
-    /// Export the dumped txout set to a CSV file.
+    /// Path to export the UTXO set in CSV format.
     #[clap(short, long, value_name = "PATH")]
     csv: Option<PathBuf>,
 
-    /// Path to the binary dump file for the UTXO set.
+    /// Path to export the UTXO set as a binary file.
     ///
     /// The binary dump is compatible with the format used by the Bitcoin Core client.
-    /// You can export the UTXO set from Subcoin and import it into the Bitcoin
-    /// Core client.
     ///
     /// If neither `--csv` nor `--binary` options are provided, the UTXO set will be
     /// printed to stdout in CSV format.
@@ -56,20 +54,13 @@ impl DumpTxOutSetCmd {
 
         let (block_number, bitcoin_block_hash, utxo_iter) = fetch_utxo_set_at(&client, height)?;
 
-        let mut output_file = if let Some(path) = csv {
-            println!(
-                "Dumping UTXO set at #{block_number},{bitcoin_block_hash} to {}",
-                path.display()
-            );
-            UtxoSetOutput::Csv(std::fs::File::create(path)?)
-        } else if let Some(path) = binary {
+        if let Some(path) = binary {
             let utxo_set_size = fetch_utxo_set_at(&client, height)?.2.count() as u64;
 
             println!(
-                "Dumping UTXO set at #{block_number},{bitcoin_block_hash} to {}",
+                "Exporting UTXO set (utxo_set_size) at #{block_number},{bitcoin_block_hash} to binary file: {}",
                 path.display()
             );
-            println!("UTXO set size: {utxo_set_size}");
 
             let file = std::fs::File::create(&path)?;
 
@@ -79,41 +70,37 @@ impl DumpTxOutSetCmd {
             snapshot_generator.generate_snapshot_in_mem(
                 bitcoin_block_hash,
                 utxo_set_size,
-                utxo_iter.map(|(txid, vout, coin)| subcoin_utxo_snapshot::Utxo {
-                    txid,
-                    vout,
-                    coin,
-                }),
+                utxo_iter.map(Into::into),
             )?;
 
             return Ok(());
+        }
+
+        let mut output_file = if let Some(path) = csv {
+            println!(
+                "Exporting UTXO set at #{block_number},{bitcoin_block_hash} to CSV file: {}",
+                path.display()
+            );
+            UtxoSetOutput::Csv(std::fs::File::create(path)?)
         } else {
-            println!("Dumping UTXO set at #{block_number},{bitcoin_block_hash}");
+            println!("Exporting UTXO set at #{block_number},{bitcoin_block_hash} to stdout:");
             UtxoSetOutput::Stdout(std::io::stdout())
         };
 
         let processed = Arc::new(AtomicUsize::new(0));
 
-        let ordered_utxos = subcoin_utxo_snapshot::group_utxos_by_txid(
-            utxo_iter
-                .into_iter()
-                .map(|(txid, vout, coin)| subcoin_utxo_snapshot::Utxo { txid, vout, coin }),
+        let grouped_utxos =
+            subcoin_utxo_snapshot::group_utxos_by_txid(utxo_iter.into_iter().map(Into::into));
+
+        let total_txids = grouped_utxos.len();
+
+        crate::utils::show_progress_in_background(
+            processed.clone(),
+            total_txids as u64,
+            format!("Dumping UTXO set at block #{block_number}..."),
         );
 
-        let total_txids = ordered_utxos.len();
-
-        std::thread::spawn({
-            let processed = processed.clone();
-            move || {
-                super::get_txout_set_info::show_progress(
-                    processed,
-                    total_txids as u64,
-                    format!("Dumping UTXO set at block #{block_number}..."),
-                )
-            }
-        });
-
-        for (txid, coins) in ordered_utxos {
+        for (txid, coins) in grouped_utxos {
             for subcoin_utxo_snapshot::OutputEntry { vout, coin } in coins {
                 output_file.write(txid, vout, coin)?;
             }
@@ -127,7 +114,6 @@ impl DumpTxOutSetCmd {
 }
 
 enum UtxoSetOutput {
-    Snapshot(UtxoSnapshotGenerator),
     Csv(File),
     Stdout(Stdout),
 }
@@ -135,9 +121,6 @@ enum UtxoSetOutput {
 impl UtxoSetOutput {
     fn write(&mut self, txid: bitcoin::Txid, vout: u32, coin: Coin) -> std::io::Result<()> {
         match self {
-            Self::Snapshot(snapshot_generator) => {
-                snapshot_generator.write_utxo_entry(txid, vout, coin)?;
-            }
             Self::Csv(ref mut file) => {
                 let Coin {
                     is_coinbase,
